% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_haz_src.R
\name{examine_haz_src}
\alias{examine_haz_src}
\title{Examine the FORTRAN file matches the format or not}
\usage{
examine_haz_src(fault.src, fault_output)
}
\arguments{
\item{fault.src}{character: File name of 'fault.src'}

\item{fault_output}{character: New file name of 'fault.src'}
}
\value{
new FORTRAN file of \code{fault.src}
}
\description{
Examine the FORTRAN file matches the format or not
}
