% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_information.R
\name{fig_information}
\alias{fig_information}
\title{Figure Information}
\usage{
fig_information(tabfile, outputname)
}
\arguments{
\item{tabfile}{character string: File name of figure (.TAB)}

\item{outputname}{character string: Set the new output name}
}
\value{
table of \code{tabfile}
}
\description{
Figure Information
}
