% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figinfo.R
\name{figinfo}
\alias{figinfo}
\title{Figure Information}
\usage{
figinfo(tabfile, outputname)
}
\arguments{
\item{tabfile}{character string: File name of figure (.TAB)}

\item{outputname}{character string: Set the new output name}
}
\value{
table of \code{tabfile}
}
\description{
Figure Information
}
