% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fort2table.R
\name{fort2table}
\alias{fort2table}
\title{Fortran to Table}
\usage{
fort2table(faultfile)
}
\arguments{
\item{faultfile}{character string: File name of 'fault.src'}
}
\value{
table of \code{faultfile}
}
\description{
Fortran to Table
}
