% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fault_table.R
\name{output_fault_table}
\alias{output_fault_table}
\title{Table output}
\usage{
output_fault_table(path, faultfile)
}
\arguments{
\item{path}{character string : File path. Including all of additional file.}

\item{faultfile}{character string : File name of 'fault.src'}
}
\value{
FName1 : Fault Name.

probAct0 : Probability of activity.

nSeg Model : Number of segment model.

segWt : Weights for each segment model.

nFlt2 : Number of segment.

Fault ID : ID number for each fault. This value will be use in \code{output_table}
}
\description{
Table of all fault segements,including Fault Name, probAct0, nSeg Model, segWt, nFlt2 and Fault ID.
}
