% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_segment_table.R
\name{output_segment_table}
\alias{output_segment_table}
\title{Table output}
\usage{
output_segment_table(path, faultfile, ID)
}
\arguments{
\item{path}{character string: File path. Including all of additional file.}

\item{faultfile}{character string: File name of 'fault.src'}

\item{ID}{integer: fault ID}
}
\value{
table of \code{faultfile}
}
\description{
Table output
}
