% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_map.R
\name{plot_all_map}
\alias{plot_all_map}
\title{Plot map of all fault segments}
\usage{
plot_all_map(faultfile)
}
\arguments{
\item{faultfile}{list: result of \code{\link{read_fault}}}
}
\value{
leaflet map
}
\description{
Plot map of all fault segments
}
