% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haz_gis.R
\name{plot_haz_gis}
\alias{plot_haz_gis}
\title{Plot HAZ GIS}
\usage{
plot_haz_gis(fault_gis)
}
\arguments{
\item{fault_gis}{character: Result of \code{read_haz_gis()}}
}
\value{
figure of \code{fault_gis}
}
\description{
Plot HAZ GIS
}
