% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haz_src_all.R
\name{plot_haz_src_all}
\alias{plot_haz_src_all}
\title{Plot map of all fault segments}
\usage{
plot_haz_src_all(faults)
}
\arguments{
\item{faults}{list: result of \code{\link{read_haz_src}}}
}
\value{
leaflet map
}
\description{
Plot map of all fault segments
}
