% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haz_src_segment.R
\name{plot_haz_src_segment}
\alias{plot_haz_src_segment}
\title{Plot map of specific fault segment}
\usage{
plot_haz_src_segment(faults, ID)
}
\arguments{
\item{faults}{list: result of \code{\link{read_haz_src}}}

\item{ID}{intrger: fault ID}
}
\value{
leaflet map
}
\description{
Plot map of specific fault segment
}
