% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segment.R
\name{plot_segment}
\alias{plot_segment}
\title{Plot map of specific fault segment}
\usage{
plot_segment(faultfile, ID)
}
\arguments{
\item{faultfile}{list: result of \code{\link{read_fault}}}

\item{ID}{intrger: fault ID}
}
\value{
leaflet map
}
\description{
Plot map of specific fault segment
}
