% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fault.R
\name{read_fault}
\alias{read_fault}
\title{Read file about fault.src}
\usage{
read_fault(path, faultfile)
}
\arguments{
\item{path}{character string : File path. Including all of additional file.}

\item{faultfile}{character string : File name of 'fault.src'}
}
\value{
a list of \code{faultfile}
}
\description{
Read file about fault.src
}
