% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_haz_gis.R
\name{read_haz_gis}
\alias{read_haz_gis}
\title{Read HAZ GIS}
\usage{
read_haz_gis(fault_tab)
}
\arguments{
\item{fault_tab}{character string: File name of figure (.TAB)}
}
\value{
table of \code{fault_tab}
}
\description{
Read HAZ GIS
}
