% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_haz_src.R
\name{read_haz_src}
\alias{read_haz_src}
\title{Fortran to Table}
\usage{
read_haz_src(fault.src)
}
\arguments{
\item{fault.src}{character: File name of 'fault.src'}
}
\value{
table of \code{fault.src}
}
\description{
Fortran to Table
}
