% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read file about input.inp}
\usage{
read_input(input.inp)
}
\arguments{
\item{input.inp}{character string : input.inp file}
}
\value{
a list of \code{input.inp}
}
\description{
Read file about input.inp
}
