% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2fort.R
\name{table2fort}
\alias{table2fort}
\title{Table to Fortran}
\usage{
table2fort(icoor, inputfile, outputname)
}
\arguments{
\item{icoor}{numeric: The unit of the source coordinate, 0 is km, 1 is latitude and longitude}

\item{inputfile}{character string: Table of 'fault.src'}

\item{outputname}{character string: Set the new output name}
}
\value{
fortran file of \code{faultfile}
}
\description{
Table to Fortran
}
