% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haz_gis.R
\name{write_haz_gis}
\alias{write_haz_gis}
\title{Figure Information}
\usage{
write_haz_gis(fault_gis, fault_gis_out, driver = "MapInfo File")
}
\arguments{
\item{fault_gis}{character: Result of \code{read_haz_gis()}}

\item{fault_gis_out}{character: Output file name.}

\item{driver}{character: Output driver. Use \code{st_drivers()} to get more information; default: \code{FALSE}}
}
\value{
shapefile of \code{fault_gis}
}
\description{
Figure Information
}
